Update build.py
 - Fixed essentials rename bug. The file structure displayed in Pho's README.md is incorrect. A constant variable has been added to control whether the program modifies file name to match Pho's file naming standard.

+++ build.py
@@ -6,6 +6,10 @@ import pathlib
 from PIL import Image
 import shutil
 from datetime import datetime
+import sys
+from PIL.ExifTags import TAGS
+from win32com.propsys import propsys, pscon
+import win32timezone
 
 
 # 判断文件是否在某一文件夹以下
@@ -57,18 +61,32 @@ def process_photos(folder_path, target_folder):
                 continue
 
             # 使用Pillow库打开图片并获取拍摄日期
+            date_time: datetime = datetime.fromtimestamp(os.path.getmtime(item_path))
+
+            # 图片类型 
+            # 读取 EXIF 中的“Date Taken”参数，参考
+            # https://stackoverflow.com/questions/23064549/get-date-and-time-when-photo-was-taken-from-exif-data-using-pil
             try:
                 with Image.open(item_path) as img:
                     exif_data = img._getexif()
-                    # 根据exif中的拍摄日期信息获取日期
-                    if 36867 in exif_data:
-                        date_str = exif_data[36867].strip().replace(" ", "")
-                        # print("info:", date_str)
-                        date_time = datetime.strptime(date_str, "%Y:%m:%d%H:%M:%S").date()
-                    else:
-                        date_time = datetime.fromtimestamp(os.path.getmtime(item_path)).date()
-            except (AttributeError, KeyError, IndexError, OSError, TypeError):
-                date_time = datetime.fromtimestamp(os.path.getmtime(item_path)).date()
+                    # Pho 读取的是 12 小时制的 Date Taken，而不是24小时制！
+                    # date_time = datetime.strptime(exif_data[36867], '%Y:%m:%d %H:%M:%S')
+                    time24 = datetime.strptime(exif_data[36867], '%Y:%m:%d %H:%M:%S')
+                    time12 = datetime.strftime(time24, '%Y:%m:%d %I:%M:%S')
+                    date_time = datetime.strptime(time12, '%Y:%m:%d %H:%M:%S')
+            except (AttributeError, KeyError, IndexError, OSError, TypeError) as e:
+                pass
+
+            # 视频类型 读取 Media created 属性，Pho 读取的就是这个属性 参考 
+            # https://stackoverflow.com/questions/79613425/get-media-created-timestamp-with-python-for-mp4-and-m4a-video-audio-files
+            try:
+                properties = propsys.SHGetPropertyStoreFromParsingName(item_path)
+                dt = properties.GetValue(pscon.PKEY_Media_DateEncoded).GetValue()
+                date_time = datetime.fromtimestamp(dt.timestamp())
+            except (AttributeError, KeyError, IndexError, OSError, TypeError) as e:
+                pass
+            finally:
+                del dt, properties  # 不释放无法 Move，文件被占用
 
             # 创建目标文件夹，如果不存在
             target_date_folder = os.path.join(target_folder, str(date_time.year), str(date_time.month).zfill(2),
@@ -78,8 +96,18 @@ def process_photos(folder_path, target_folder):
 
             # 移动文件到目标文件夹
             dest_path = os.path.join(target_date_folder, item)
-            shutil.move(item_path, dest_path)
-            print('[%s%%] Moved %s' % (int(file_count/total_files * 100), dest_path))
+
+            # 使用 Pho 命名格式 20250215082557_AAA.jpg
+            if use_pho_standard:
+                item = f'{str(date_time.year)}{str(date_time.month).zfill(2)}{str(date_time.day).zfill(2)}{str(date_time.hour).zfill(2)}{str(date_time.minute).zfill(2)}{str(date_time.second).zfill(2)}_' + item
+                dest_path = os.path.join(target_date_folder, item)
+
+            # 如果已存在，则不移动 
+            if os.path.exists(dest_path):
+                print('[%s%%] Exists cancel moving %s' % (int(file_count / total_files * 100), dest_path))
+            else:
+                shutil.move(item_path, dest_path)
+                print('[%s%%] Moved %s' % (int(file_count / total_files * 100), dest_path))
 
 
 # 获取自身文件路径
@@ -92,13 +120,20 @@ def self_path():
         return os.path.abspath(os.path.dirname(__file__))
 
 
+# Pho 命名规则
+print(
+    '注：Pho 的命名规则要求在文件名称前增添时间戳，这与其公开目录结构不符，应该为 Bug。如要复原，在 APP 内随意上传一个文件就可以知道。')
+print('名称示例：AAA.mp4 -> 20250215082557_AAA.mp4')
+use_pho_standard = True
+
 # 定义支持的文件
 supported_extensions = ['.png', '.jpg', '.jpeg', '.gif', '.bmp', '.heic', '.mp4', '.dng']
 
 # 定义照片文件夹和目标文件夹
 photo_folder = self_path()  # 该py脚本所在路径
 target_folder = "%s\\Completed" % photo_folder  # 该py脚本所在路径下的Completed文件夹
-print('处理目录：%s' % target_folder)
+print('输入目录：%s' % photo_folder)
+print('输出目录：%s' % target_folder)
 
 # 解除Pillow默认最大像素限制
 Image.MAX_IMAGE_PIXELS = None
